backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
beautifulsoup4==4.14.2
requests==2.32.5
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
numpy==2.3.3
weaviate-client==3.1.0



